# Speed up timestamp generation when logging