This PR speeds up how we generate timestamps when logging; instead of regenerating them from scratch all the time we now cache part of the timestamp in TLS.

Here's the performance of the original implementation as measured on my machine:

```
chrono_local            time:   [871.11 ns 872.19 ns 873.30 ns] 
```

And here's the performance of the new one:

```
fast_local_time         time:   [35.994 ns 36.011 ns 36.030 ns]
```

So the overhead was cut down to less than 10%.

## Seems like this was pretty cheap in the first place; does this actually matter?

This will *not* help in normal cases where we're not logging much, but anytime someone turns on a spammy trace log (which does happen on our staking ops' nodes) the costs of logging starts to balloon. Yes, we shouldn't be logging this much. But if we do then it shouldn't bog down the CPU as much as it does.

Even though this might seem like it shouldn't make much difference the issue here is that at this scale we're experiencing a case of a death by a thousand cuts, where even though doing it only *once* is very cheap we're doing it so much that once you add everything together you end up in a situation where logging can take a whopping 30% of the total CPU time. This is one of the easier and most straightforward improvements we can make, which represented ~6.6% of the total CPU used in this particular profiling run, which now should be cut down to less than ~1%.